Rename workflow modules for concise imports
